handle all Exception in task and employee

+++ employee/src/main/java/com/ebi/employee/service/EmployeeServiceImplementation.java
@@ -1,6 +1,8 @@
 package com.ebi.employee.service;
 
-import com.ebi.employee.exception.CustomException;
+import com.ebi.employee.exception.EmployeeGetException;
+import com.ebi.employee.exception.EmployeeSaveException;
+import com.ebi.employee.exception.EmployeeUpdateException;
 import com.ebi.employee.model.EmployeeDto;
 import com.ebi.employee.entity.EmployeeEntity;
 import com.ebi.employee.model.EmployeeSaveDto;
@@ -35,12 +37,25 @@ public class EmployeeServiceImplementation implements  EmployeeServiceInterface{
         if(employeeEntity.isPresent())
             return modelMapper.map(employeeEntity.get(),EmployeeDto.class);
         else
-            throw new CustomException("888","Not Found","No element of data has this ID : "+id) ;
+            throw new EmployeeGetException("01","Not Found","No element of data has this ID : " + id) ;
+    }
+
+    @Override
+    public EmployeeDto getEmployeeByNameAndMail(String name, String email) {
+        Optional<EmployeeEntity> employeeEntity = employeeRepoInterface.findByMyQuery(name,email);
+        if(employeeEntity.isPresent())
+            return modelMapper.map(employeeEntity.get(),EmployeeDto.class);
+        else
+            throw new EmployeeGetException("051","Miss Employee ","No Employee with name : "+name+" and email : "+email);
     }
 
     @Override
     public EmployeeDto saveEmployee(EmployeeSaveDto employee){
         EmployeeEntity employeeEntity = modelMapper.map(employee,EmployeeEntity.class);
+        if(employeeEntity.getPhone().length()!=11||!employeeEntity.getPhone().startsWith("01"))
+            throw new EmployeeSaveException("021","Phone Number Error","Phone number must be 11 digit and start with 01");
+        if(!employeeEntity.getEmail().contains(".com")||!employeeEntity.getEmail().contains("@"))
+            throw new EmployeeSaveException("022","Email Address Error","Email Address contains @ and .com");
         employeeRepoInterface.save(employeeEntity);
         return modelMapper.map(employeeEntity,EmployeeDto.class);
     }
@@ -48,6 +63,12 @@ public class EmployeeServiceImplementation implements  EmployeeServiceInterface{
     @Override
     public EmployeeDto updateEmployee(EmployeeSaveDto employee){
         EmployeeEntity employeeEntity =  modelMapper.map(employee,EmployeeEntity.class);
+        if(employeeEntity.getEmail()==null||employeeEntity.getSalary()==null||employeeEntity.getAddress()==null||employeeEntity.getPhone()==null||employeeEntity.getName()==null)
+            throw new EmployeeUpdateException("033","Empty Field","Must Enter Name , Phone , Email , Salary , Address .");
+        if(employeeEntity.getPhone().length()!=11||!employeeEntity.getPhone().startsWith("01"))
+            throw new EmployeeUpdateException("031","Phone Number Error","Phone number must be 11 digit and start with 01");
+        if(!employeeEntity.getEmail().contains(".com")||!employeeEntity.getEmail().contains("@"))
+            throw new EmployeeUpdateException("032","Email Address Error","Email Address contains @ and .com");
         EmployeeEntity employeeEntity1 = employeeRepoInterface.save(employeeEntity);
         return modelMapper.map(employeeEntity1,EmployeeDto.class);
     }
@@ -61,40 +82,58 @@ public class EmployeeServiceImplementation implements  EmployeeServiceInterface{
             Optional<EmployeeEntity> employeeEntityOptional = employeeRepoInterface.findById(employee.getId());
         if(employeeEntityOptional.isPresent()){
             if (employee.getName()!=null) {
-                employeeEntityOptional.get().setName(employee.getName());
+               if(!employee.getName().equals(employeeEntityOptional.get().getName()))
+                   employeeEntityOptional.get().setName(employee.getName());
+               else
+                   throw new EmployeeUpdateException("043","Name Error","Must Enter New Name ");
             }
             if (employee.getSalary()!=null) {
-                employeeEntityOptional.get().setSalary(employee.getSalary());
+                if(!employee.getSalary().equals(employeeEntityOptional.get().getSalary()))
+                    employeeEntityOptional.get().setSalary(employee.getSalary());
+                else
+                    throw new EmployeeUpdateException("044","Salary Error","Must Enter New Salary ");
             }
             if (employee.getAddress()!=null)
             {
-                employeeEntityOptional.get().setAddress(employee.getAddress());
+                if(!employee.getAddress().equals(employeeEntityOptional.get().getAddress()))
+                    employeeEntityOptional.get().setAddress(employee.getAddress());
+                else
+                    throw new EmployeeUpdateException("045","Address Error","Must Enter New Address ");
             }
             if (employee.getEmail()!=null)
             {
-                employeeEntityOptional.get().setEmail(employee.getEmail());
+                if(!employee.getEmail().equals(employeeEntityOptional.get().getEmail()) && employee.getEmail().contains(".com") && employee.getEmail().contains("@"))
+                    employeeEntityOptional.get().setEmail(employee.getEmail());
+                else
+                    throw new EmployeeUpdateException("046","Email Error","Must Enter New Email must include @ and .com");
             }
             if (employee.getPhone()!=null)
             {
-                employeeEntityOptional.get().setPhone(employee.getPhone());
+                if(employee.getPhone().length()==11 && employee.getPhone().startsWith("01") && !employee.getPhone().equals(employeeEntityOptional.get().getPhone()))
+                    employeeEntityOptional.get().setPhone(employee.getPhone());
+                else
+                    throw new EmployeeUpdateException("047","Phone Error","Must Enter New Phone has 11 digit and start with 01");
             }
         }else
-            throw new CustomException("700","Miss Data","Can not find element for this id : "+employee.getId());
+            throw new EmployeeGetException("042","Miss Employee","No Employee with id : "+employee.getId());
         savedEmployeeEntity = employeeRepoInterface.save(employeeEntityOptional.get());
         } else
-            throw new CustomException("777","miss Data","Can not find Resave data");
+            throw new EmployeeUpdateException("041","Miss ID","Must Send Employee ID");
         return modelMapper.map(savedEmployeeEntity,EmployeeDto.class);
     }
 
     @Override
-    public void deleteEmployee(long id){
-         employeeRepoInterface.deleteById(id);
+    public EmployeeDto deleteEmployee(Long id){
+        Optional<EmployeeEntity> employeeEntityOptional = employeeRepoInterface.findById(id);
+        if(employeeEntityOptional.isPresent())
+        {
+            employeeRepoInterface.deleteById(id);
+            return modelMapper.map(employeeEntityOptional.get(),EmployeeDto.class);
+        }
+        else
+            throw new EmployeeUpdateException("061","Miss Employee","No Employee with id : "+id+" to Delete it .");
     }
 
-    @Override
-    public EmployeeDto getEmployeeByNameAndMail(String name, String email) {
-        Optional<EmployeeEntity> employeeEntity = employeeRepoInterface.findByMyQuery(name,email);
-        return employeeEntity.map(entity -> modelMapper.map(entity, EmployeeDto.class)).orElse(null);
-    }
+
 
 }
